# meta: improve CheckSpecialAttributes