… pkg/meta/BUILD.bazel