<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #50959

Problem Summary:

### What changed and how does it work?

I see it's the bottleneck, and we don't need to scan the whole string for 4 times to check 4 substring.

![image](https://github.com/user-attachments/assets/3477cdd0-43fc-4bb8-a09c-7768e35986ab)

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->


before

```
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/meta
BenchmarkCheckSpecialAttributes
BenchmarkCheckSpecialAttributes/narrow
BenchmarkCheckSpecialAttributes/narrow-8         	  637102	      1845 ns/op
BenchmarkCheckSpecialAttributes/wide
BenchmarkCheckSpecialAttributes/wide-8           	  105760	     11328 ns/op
```

after 

```
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/meta
BenchmarkCheckSpecialAttributes
BenchmarkCheckSpecialAttributes/narrow
BenchmarkCheckSpecialAttributes/narrow-8         	  957866	      1287 ns/op
BenchmarkCheckSpecialAttributes/wide
BenchmarkCheckSpecialAttributes/wide-8           	  326424	      3674 ns/op
```

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```